Refactor Devsktop component structure and introduce WindowMenuWrapper
- Updated index.ts to streamline exports for WindowMenuWrapper and other components.
- Created WindowMenuWrapper component to manage menu overlays for window components, enhancing modularity and usability.
- Refactored devsktop-icons and window-frame components to improve code clarity and maintainability.
- Removed unused props interfaces and adjusted related components for consistency.
