+++ src/features/devsktop/components/devsktop-icons/devsktop-icons.tsx
@@ -15,7 +15,7 @@ function useWindowDimensions() {
 		};
 
 		window.addEventListener('resize', handleResize);
-		return () => window.removeEventListener('resize', handleResize);
+		return () => { window.removeEventListener('resize', handleResize) };
 	}, []);
 
 	return dimensions;
