+++ src/features/devsktop/components/index.ts
@@ -1,7 +1,7 @@
 export { NavigationBar } from './navigation-bar/navigation-bar';
 export { WindowFrame } from './window-frame/window-frame';
 export { DevsktopIcons } from './devsktop-icons/devsktop-icons';
-export { WindowMenuWrapper } from './window-menu-wrapper/window-menu-wrapper';
+export { WindowMenuWrapper } from './window-menu-wrapper';
 export { LandingPageIntro } from './landing-page/landing-page-intro';
 export { LanguagesGrid } from './landing-page/languages-grid';
 export { FrameworksGrid } from './landing-page/frameworks-grid';
