+++ src/features/devsktop/components/window-menu-wrapper.tsx
@@ -1,5 +1,17 @@
-import { CloseIcon } from '../../../../assets';
-import type { WindowMenuWrapperProps } from './window-menu-wrapper-props';
+/* eslint-disable @typescript-eslint/no-explicit-any */
+import type { Dispatch, ReactNode, SetStateAction } from 'react';
+import { CloseIcon } from '../../../assets';
+import type { IconProps, ManagedWindow, SiteSettings } from '../../../types';
+
+export interface WindowMenuWrapperProps {
+	siteSettings: SiteSettings;
+	windowState: ManagedWindow;
+	showMenu: boolean;
+	setShowMenu: Dispatch<SetStateAction<boolean>>;
+	menuMainTitle: string;
+	Content: (props: any) => ReactNode;
+	Icon: (props: IconProps) => ReactNode;
+}
 
 /**
  * A wrapper component that renders a menu overlay for window components.
