+++ src/features/solvedoku/components/cell-data.tsx
@@ -5,8 +5,28 @@ import {
 	TailwindBGs300,
 	TailwindOutlineColors,
 	TailwindTextColors500,
-} from '../../../../functions';
-import type { CellDataProps } from './cell-data-props';
+} from '../../../functions';
+import type { Cell, Colors } from '../../../types';
+
+
+export interface CellDataProps {
+	cellID: string;
+	rowIndex: number;
+	colIndex: number;
+	cell: Cell;
+	cellSizing: string;
+	accentColor: Colors;
+	selectedCellCoords: [number, number] | null;
+	selectedCellID: string | null;
+	updateCellValue: (
+		cellID: string,
+		value: string | null,
+		userInput: boolean
+	) => void;
+	setSelectedCellID: (cellID: string | null) => void;
+	showingStoredSolution: boolean;
+}
+
 
 /**
  * Cell component for the Sudoku game board.
@@ -93,22 +113,22 @@ export const CellData = React.memo(function CellData({
 				'outline-3 -outline-offset-4 '.concat(
 					TailwindOutlineColors[accentColor]
 				)
-			:	'border border-black';
+				: 'border border-black';
 
 		return selectionOutline + (hasThickRightBorder ? ' border-r-4' : '');
 	};
 
 	const backgroundColor =
 		isSelected ? TailwindBGs300[accentColor]
-		: isRelatedToSelected ? 'bg-yellow-50'
-		: 'bg-neutral-50';
+			: isRelatedToSelected ? 'bg-yellow-50'
+				: 'bg-neutral-50';
 
 	const textColor =
 		showingStoredSolution ?
 			!cell.userInputted ?
 				TailwindTextColors500[accentColor]
-			:	'text-black'
-		:	'text-black';
+				: 'text-black'
+			: 'text-black';
 
 	return (
 		<td
