+++ src/features/solvedoku/components/puzzle-button.tsx
@@ -1,5 +1,15 @@
-import { TailwindBGs500 } from '../../../../functions';
-import type { PuzzleButtonProps } from './puzzle-button-props';
+import type { ReactNode } from 'react';
+import { TailwindBGs500 } from '../../../functions';
+import type { Colors, IconProps } from '../../../types';
+
+export interface PuzzleButtonProps {
+	title: string;
+	Icon?: ({ size, classes }: IconProps) => ReactNode;
+	text?: string;
+	onClickFunction: () => void;
+	isDisabled: boolean;
+	accentColor: Colors;
+}
 
 /**
  * A reusable button component for puzzle-related actions
@@ -32,7 +42,7 @@ export function PuzzleButton({
 			className={`border border-gray-300/50 rounded px-1 tracking-tighter font-semibold ${bgColorClass} xs:text-xl sm:text-3xl text-nowrap transition-all ${isDisabled ? '' : 'duration-100 hover:-translate-y-1 hover:scale-105'}`}>
 			{Icon ?
 				<Icon size='w-6 xs:w-9 sm:w-12' />
-			:	<p>{text}</p>}
+				: <p>{text}</p>}
 		</button>
 	);
 }
